Move slate providers into a separate directory
